full cuda

+++ src/raw2film/raw_conversion.py
@@ -1,11 +1,10 @@
 import cv2 as cv
 import ffmpeg
 import rawpy
+from raw2film import effects
+from raw2film.color_processing import calc_exposure, xyz_to_srgb, xyz_to_displayP3
 from spectral_film_lut.film_spectral import FilmSpectral
 from spectral_film_lut.utils import *
-
-from raw2film import effects
-from raw2film.color_processing import calc_exposure
 
 
 def raw_to_linear(src, half_size=True):
@@ -42,20 +41,22 @@ def process_image(image, negative_film, grain_size, frame_width=36, frame_height
         kwargs["exp_comp"] = exp_comp
 
     if fast_mode:
-        if image.dtype != np.uint16:
-            image = image.astype(np.float32)
+        if image.dtype != xp.uint16 and not cuda_available:
+            image = image.astype(xp.float32)
             image_max = image.max()
             if image_max > 65535:
                 factor = 65535. / image_max
                 image *= factor
                 if "exp_comp" in kwargs:
-                    kwargs["exp_comp"] -= np.log2(factor)
+                    kwargs["exp_comp"] -= xp.log2(factor)
                 else:
-                    kwargs["exp_comp"] = np.log2(factor)
-            image = image.astype(np.uint16)
+                    kwargs["exp_comp"] = xp.log2(factor)
+            image = image.astype(xp.uint16)
+        elif cuda_available:
+            image = image.astype(xp.float32) / 65535
         mode = 'full'
     else:
-        image = image.astype(np.float32) / 65535
+        image = image.astype(xp.float32) / 65535
         mode = 'print'
 
     if not fast_mode:
@@ -74,9 +75,10 @@ def process_image(image, negative_film, grain_size, frame_width=36, frame_height
         h //= 2
         w //= 2
         image = cv.resize(image, (w, h), interpolation=cv.INTER_AREA)
-        image = ((image / 65535) ** 0.25 * 65535).astype(np.uint16)
-        kwargs["gamma"] = 4
-        kwargs["lut_size"] = 17
+        if not cuda_available:
+            image = ((image / 65535) ** 0.25 * 65535).astype(xp.uint16)
+            kwargs["gamma"] = 4
+            kwargs["lut_size"] = 17
     else:
         image = xp.asarray(image)
         scale = max(image.shape) / max(frame_width, frame_height)  # pixels per mm
@@ -102,29 +104,40 @@ def process_image(image, negative_film, grain_size, frame_width=36, frame_height
             if measure_time:
                 print(f"{'grain':28} {time.time() - start_sub:.4f}s {image.dtype} {image.shape} {type(image)}")
 
-        image = xp.clip(image, 0, 1)
-        image *= 2 ** 16 - 1
-        image = to_numpy(image.astype(np.uint16))
-
-    lut = create_lut(negative_film, print_film, name=str(time.time()), mode=mode, input_colourspace=None, **kwargs)
+        if not cuda_available:
+            image = xp.clip(image, 0, 1)
+            image *= 2 ** 16 - 1
+            image = to_numpy(image.astype(xp.uint16))
 
     start_sub = time.time()
-    if image.shape[-1] == 1:
-        image = image.repeat(3, -1)
+    if cuda_available:
+        if "output_colourspace" in kwargs and kwargs["output_colourspace"] == "Display P3":
+            output_transform = xyz_to_displayP3
+        else:
+            output_transform = xyz_to_srgb
+        transform, _ = FilmSpectral.generate_conversion(negative_film, print_film, mode=mode, input_colourspace=None,
+                                                        output_transform=output_transform, matrix_method=False, **kwargs)
+        image = transform(image)
+    else:
+        lut = create_lut(negative_film, print_film, name=str(time.time()), mode=mode, input_colourspace=None, **kwargs)
 
-    height, width, _ = image.shape
-    process = run_async(
-        ffmpeg.input('pipe:', format='rawvideo', pix_fmt='rgb48', s='{}x{}'.format(width, height)).filter('lut3d',
-                                                                                                          file=lut).output(
-            'pipe:', format='rawvideo', pix_fmt='rgb24', vframes=1, loglevel='quiet'), pipe_stdin=True,
-        pipe_stdout=True)
-    process.stdin.write(image.tobytes())
-    process.stdin.close()
-    image = process.stdout.read(width * height * 3)
-    process.wait()
-    os.remove(lut)
+        if image.shape[-1] == 1:
+            image = image.repeat(3, -1)
+
+        height, width, _ = image.shape
+        process = run_async(
+            ffmpeg.input('pipe:', format='rawvideo', pix_fmt='rgb48', s='{}x{}'.format(width, height)).filter('lut3d',
+                                                                                                              file=lut).output(
+                'pipe:', format='rawvideo', pix_fmt='rgb24', vframes=1, loglevel='quiet'), pipe_stdin=True,
+            pipe_stdout=True)
+        process.stdin.write(image.tobytes())
+        process.stdin.close()
+        image = process.stdout.read(width * height * 3)
+        process.wait()
+        os.remove(lut)
+        image = np.frombuffer(image, np.uint8).reshape([height, width, 3])
     if measure_time:
         print(f"{'lut':28} {time.time() - start_sub:.4f}s")
         print(f"{'total':28} {time.time() - start:.4f}s")
 
-    return np.frombuffer(image, np.uint8).reshape([height, width, 3])
+    return image
